Testshape

+++ ProteinPairsGenerator/datasets/sabdab_dataset.py
@@ -47,7 +47,7 @@ def cdr_extracter(data_pdb: AtomGroup, Lchain: List[str] = [], Hchain: List[str]
         edge_index, edge_attr = remove_self_loops(edge_index, edge_attr)
 
         # Create data point
-        seq = seq.type(torch.DoubleTensor)
+        seq = seq.type(torch.LongTensor)
         edge_attr = edge_attr.type(torch.DoubleTensor)
         edge_index = edge_index.type(torch.LongTensor)
